Emissive particles and renderPipeline
Fixed an issue where an odd number of post effects returned alternating texture id's (as it should) which caused a terrible latency or frame swapping issue with the vr headset, mirror window still looked fine though...
Fixed now

+++ src/com/theincgi/lwjglApp/Utils.java
@@ -124,7 +124,7 @@ public class Utils {
 	}
 	
 	/**Returns a normaly distributed float between -range and range, non inclusive*/
-	public static float ndRandom(int range) {
+	public static float ndRandom(float range) {
 		return (float)( Math.random()+Math.random()+Math.random() - Math.random() - Math.random() - Math.random() )/6*range;
 	}
 }
